Temporary fix to editing window crash
Just found out that attempting to edit imported resources will crash this app, and it's because imported resources have no set type and EditWindow requires a type value. The current fix is to set a default value for now, but I'll investigate if I could retrieve the type from import.